refactor(torii-grpc): event subscription with multiple clauses
