+++ crates/torii/grpc/src/server/subscriptions/event.rs
@@ -19,16 +19,17 @@ use torii_core::sql::FELT_DELIMITER;
 use torii_core::types::Event;
 use tracing::{error, trace};
 
+use super::match_keys;
 use crate::proto;
 use crate::proto::world::SubscribeEventsResponse;
-use crate::types::{KeysClause, PatternMatching};
+use crate::types::EntityKeysClause;
 
 pub(crate) const LOG_TARGET: &str = "torii::grpc::server::subscriptions::event";
 
 #[derive(Debug)]
 pub struct EventSubscriber {
     /// Event keys that the subscriber is interested in
-    keys: KeysClause,
+    keys: Vec<EntityKeysClause>,
     /// The channel to send the response back to the subscriber.
     sender: Sender<Result<proto::world::SubscribeEventsResponse, tonic::Status>>,
 }
@@ -41,7 +42,7 @@ pub struct EventManager {
 impl EventManager {
     pub async fn add_subscriber(
         &self,
-        keys: KeysClause,
+        keys: Vec<EntityKeysClause>,
     ) -> Result<Receiver<Result<proto::world::SubscribeEventsResponse, tonic::Status>>, Error> {
         let id = rand::thread_rng().gen::<usize>();
         let (sender, receiver) = channel(1);
@@ -108,33 +109,7 @@ impl Service {
             .map_err(ParseError::from)?;
 
         for (idx, sub) in subs.subscribers.read().await.iter() {
-            // if the key pattern doesnt match our subscribers key pattern, skip
-            // ["", "0x0"] would match with keys ["0x...", "0x0", ...]
-            if sub.keys.pattern_matching == PatternMatching::FixedLen
-                && keys.len() != sub.keys.keys.len()
-            {
-                continue;
-            }
-
-            if !keys.iter().enumerate().all(|(idx, key)| {
-                // this is going to be None if our key pattern overflows the subscriber key pattern
-                // in this case we might want to list all events with the same
-                // key selector so we can match them all
-                let sub_key = sub.keys.keys.get(idx);
-
-                // if we have a key in the subscriber, it must match the key in the event
-                // unless its empty, which is a wildcard
-                match sub_key {
-                    // the key in the subscriber must match the key of the entity
-                    // athis index
-                    Some(Some(sub_key)) => key == sub_key,
-                    // otherwise, if we have no key we should automatically match.
-                    // or.. we overflowed the subscriber key pattern
-                    // but we're in VariableLen pattern matching
-                    // so we should match all next keys
-                    _ => true,
-                }
-            }) {
+            if !match_keys(&keys, &sub.keys) {
                 continue;
             }
